Rename lilbee.platform -> lilbee.system; drop redundant Nuitka YAML override
Nuitka's compile-time import resolver shadowed stdlib `platform` with the
sibling `lilbee.platform` module under standalone bundling, so attr/_compat.py's
`platform.python_implementation()` call hit AttributeError at runtime. Renaming
to `lilbee.system` removes the stdlib name collision and makes the import path
unambiguous for any bundler (Nuitka, PyInstaller, future).

Also drop nuitka-package-overrides/llama_cpp.yml: Nuitka 4.x's standard package
config already handles llama_cpp's lib/ directory via dlls + relative_path
(checksum 34acaa89). Our override was redundant and tripped a schema validation.

Add --no-deployment-flag=self-execution to the Nuitka invocation so subprocess
re-invocation patterns (multiprocessing spawn, etc.) don't fatally bail in
deploy mode.
