--- nuitka-package-overrides/llama_cpp.yml
@@ -1,4 +1,0 @@
-- module-name: 'llama_cpp'
-  data-files:
-    dirs:
-      - 'lib'
