+++ src/lilbee/config.py
@@ -960,7 +960,7 @@ class Config(BaseSettings):
     @model_validator(mode="before")
     @classmethod
     def _resolve_defaults(cls, data: Any) -> Any:
-        from lilbee.platform import canonical_models_dir, default_data_dir, find_local_root
+        from lilbee.system import canonical_models_dir, default_data_dir, find_local_root
 
         if not isinstance(data, dict):  # pragma: no cover
             return data
@@ -995,7 +995,7 @@ class Config(BaseSettings):
         dotenv_settings: Any,
         file_secret_settings: Any,
     ) -> tuple[Any, ...]:
-        from lilbee.platform import default_data_dir, find_local_root
+        from lilbee.system import default_data_dir, find_local_root
 
         data_env = os.environ.get("LILBEE_DATA", "")
         if data_env:
